music76489: disable reverse_byte
"reverse_byte" was only added because the protoboard had the wrong
wiring and it was easier to "fix" it in software that to re-wire the
protoboard.

Since the PCB are working Ok, there is no need to keep using the
protoboard, thus the "reverse_byte" is no longer used.

In addition to that, new code to play music from REPL was added.

And "play()" (which is not called "tick") no longer measures time.
The caller is responsible for calling it at 60Hz. This simplifies the
game logic and is less error prone.