ds3: simplify code
DS3 follows the same "pattern" used in DS4 and DS5.
Although DS3 does not use a report that is compatible with DS4 (or DS5)
for the buttons, it is good to have similar code in all of them.

This makes the code much easier to read (and mantain).

Also fixes a bug in DS3 where the player LEDs was not being set when
rumble was set at the same time.

misc: move comments to left in ds4 / ds5 making the code looks "shorter"