new platform: NINA
Add NINA as a new supported platform.
NINA and Airlift and basically the same with the exception of the MOSI
pin for SPI.
And since Airlift is a fork of NINA, to honor the original name, the
files platform_airlift_* were renamed to platform_nina_*.

Both Airlift and NINA share the same file, although they are treated as
different platforms.

Tested on Arduino Nano RP2040 Connect.
Arduino client library is still missing, coming soon, probably in its
own repo.

Also, NINA-W102 requires to have SPI in compatibility mode to work with
ESP-IDF >= v4.0. So, sdkconfig has the flag turned on by default.