NINA: code cleanup + moved commands from 0x60 to 0x00
Add more comments to code, specially in the SPI protocol.
The original Arduino code lacks some comments, and people not familiar
with it might find it confusing. Trying to solve that by adding more
comments (useful for me when I try to re-read this code in 6 months)

Moved commands from 0x60 - 0x6f to 0x00 - 0x0f.
This is because Arduino NINA are using 0x60-0x6f for file I/O commands.
Although there should be no problem to reused that space, better if we
don't reuse existing numbers... just in case.
For exmaple, the CheckNINAFirmware Sketch can be reused with out own
firmware.