ESP32 + Arduino: WIP
This commit adds Arduino as a new platform.
This is a new platform that allows Bluepad32 to use the ESP32 Arduino
Core on ESP32 modules.
With this new platform, there would be no need to use "bluepad32-nina"
platform.

This platform depends on ESP-IDF v4.4 (which is unreleased), and
requires that the "arduino-esp32" folder is added as a component in the
esp-idf repo.
For more info see:
https://docs.espressif.com/projects/arduino-esp32/en/latest/esp-idf_component.html

Additionally, the Bluetooth-Serial and BLE "submodules" needs to be
disabled from arduino-esp32.

As mentioned this is still WIP.