fw: add Kconfig option
Adds Kconfig conf file. This allows to configure Bluepad32 from:
  $ make menuconfig

This is to avoid using the configs/config-$platform file.
The config-$platform file are still going to be used. But if someone
needs to use Bluepad32 as a ESP-IDF component, it is good to remove
"ties" from the config-$platform files.

Now, when doing "make menuconfig" and option to select the Bluepad32
platform is present. By default is "none", which means that the one
defined by the PLATFORM envinroment variable will be used.

Arduino users should set it to "Arduino"