BLE: intial BLE support
BLE is needed for the new Xbox One S series gamepad.
Also, the only Bluetooth Classic mouse that I know is the obsolete Apple
mouse. But there dozens of BLE mice.
Besides that there multiple unexpensive BLE devices that can be used to
create your own HID device (think of Adafruit BLE devices).

For that reason, BLE is needed.
This commit is just the initial BLE support. Might take a while until it
is finished.