clang-format: using 120-col + 4-space indent.
In the past, the code was following the "Google" clang-format style
which seems to be good for C++ code.
But for C code, where function and variable names are longer, limiting
the col to 80, makes the code "unglier".

Also, using a 4-space indent, instead of 2... hey, since now the col is
120, we can use a bigger space indent.

The other minor change, is that the style is no longer "Google". Using
"Chromium" instead, which has better defaults.

These changes are "subjective". But since I'm spending so much time
reading the code, I want it to make it easier for me :)