make/cmake: exclude arduino files when not enabled
This commit removes Arduino files from compilation list if platform
arduino is not enabled.
In addition to that, it modifes the .c / .cpp files with and #error if
arduino platform is not enabled.

So, if an Arduino file tries to get compiled when the Arduio platform is
not selected, it will generate a compile time error.