MightyMiggy: fix alignment in comments
The original MightyMiggy code was using a different coding style than
Bluepad32 when it was original merged.
So it was converted to the 80-column, 2-spaces coding style.
But later, Bluepad32 changed the coding style to a 4-space / 120-column.

So, some of the orignal tabs were lost...
This CL fixes that by taking the original MightyMiggy code, and convert
it to the new coding style

(cherry picked from commit 8cc40ef0441774a66db9a5b670cd4f3b69659988)