uni mappings: get rid of log2()
log2() was added just as a temp solution to do a quick test.
But of course it was insante to calcualte the log 2 of an integer using
log2().

Also, the code is more consistent now.
And also it should fix the builder which was complaining that it is not
valid to initialize a struct by with log2(), which kind of makes sense.