setup_compelte, gap_inquiry with pause, and other features
This commit should have been split in two or three, but long-story
short, all the changes are somewhat related:

- DS4 supports feature reports.
- "parsers" are responsible for calling "setup_complete"
- DS3 is working again
- Wii reconnect works better.

It seems that when the gap_inquiry is on, it is more difficult for
gamepads to reconnect (Wii, DS3).
So, now the inquiry is active for 1280ms, and then there is a pause of
1000ms... during that 1000ms pause window, gamepads will have better
chances to reconnect.
-