Disable BLE, use max MTU
OMG... I was using `#ifdef ENABLE_BLE` to determine whether the BLE code
should be enabled. But you know what... btstack (or was it ESP-IDF) was
defining ENABLE_BLE.

So, release v2.5 was released with BLE enabled.
That is not a problem per-se... but the problem is that part of the BLE
code was conflicting with the Classic logic. For example, the ENCRYPTION
ROLE CHANGE... in Classic it is ignored. But used in BLE.
But it is triggered in Classic as well... and you know what... it was
causing a crash when reconnecting some devices.

Also, changes the MTU to 0xffff, which is the max.
That's what hid_host (btstack code) is doing.