uni_bluetooth: fix in device name parsing
Initialize name buffer in function.
There was a nice bug that the name was populated with the previous
discovered device because the buffer had the previous name due to
returning from the function, and entering again... and the stack was
already populated from the previous function.

funny bug.

Also, update documentation.