Add  set_event_filter HCI command
This command allows us to filter inquiry results.
We are only accepting inquiries that HID as Class of Device.

This reduces the output log... making it easier to read the log.
And perhaps it is just an ilusion, but incoming connections seems to be
accepted at the first try.

In addition to that, other misc changes were added:
 - better logs
 - updated docs