bluetooth: increase inquiry len to 3
To err on the safe side, this commit increases the inquiry len time from
two to three.
This will benefit  new connections (scanning) over re-connects (paging).

In practice, this means that "new connections" will be discovered just a
bit faster, and incoming connections a bit slower. And in some cases it
might need a frew tries.