mouse: use floats for scale
why on earth did I chose to use ints instead of floats?
Although floats are slower (much slower?), the slowdown is neglicent.
And floats give me the precision that I was looking for.
And I can use the timer to tick every 80us instead of 1us.

Would that help with power consuption? If so, great.