core: refactor gamepad with controller
"uni_gamepad_t" is deprecated, and all the functions associated with it.
Instead, "uni_controller_t" should be used instead.

This is because the controller could actually be a:
- mouse
- gamepad
- balance board
- keyboard
- etc...

So, it is easier to pass different data structures to the platforms.
And the platforms decide how to use it.

It also simplifies the code: the "updated_states" was removed since it
was not used and made the code complex.

And the battery was outside of gamepad... it is part of controller.

This code has been tested only lightly.